Fix timeout reported as cancellation through interop

+++ cs/sdk/src/Proton.Drive.Sdk/Http/HttpClientFactoryExtensions.cs
@@ -2,7 +2,7 @@
 
 internal static class HttpClientFactoryExtensions
 {
-    public static HttpClient CreateClientWithTimeout(this IHttpClientFactory httpClientFactory, int timeoutSeconds)
+    public static HttpClient CreateClientWithTimeout(this IHttpClientFactory httpClientFactory, double timeoutSeconds)
     {
         var client = httpClientFactory.CreateClient();
